requirements.txt
tensorflow-gpu==1.15
jupyter
tqdm==4.46
scikit-learn==0.22
matplotlib==3.1
